postgres

+++ db/migrate/20200303045334_add_name_to_tutors.rb
@@ -2,6 +2,6 @@ class AddNameToTutors < ActiveRecord::Migration
   def change
     add_column :tutors, :name, :string
     add_column :tutors, :experience, :integer
-    add_column :tutors, :rate, :double
+    add_column :tutors, :rate, :float
   end
 end
